requirements.txt
# Core dependencies
Flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0

# Document processing
python-docx>=0.8.11
docxtpl>=0.16.0
pypdfium2>=4.30.0
Pillow>=10.0.0

# Utilities
num2words>=0.5.12
schedule>=1.2.0
openpyxl>=3.1.0
requests>=2.31.0

# Production server
gunicorn>=21.2.0
waitress>=2.1.2
pypdfium2
